% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_k.R
\name{build_k}
\alias{build_k}
\title{Building stiffness Matrix from stiffnesses of elements
for MDOF response calculation}
\usage{
build_k(k)
}
\arguments{
\item{ks}{vector of stiffnesses of elements}
}
\description{
Building stiffness Matrix from stiffnesses of elements
for MDOF response calculation
}
