% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nigam.R
\name{nigam}
\alias{nigam}
\title{Response calculation of SDOF using Nigam & Jennings algorithm}
\usage{
nigam(y, h, w, dt)
}
\arguments{
\item{y}{absolute acc. of input ground motion}

\item{h}{damping ratio}

\item{w}{natural circular frequency (rad)}

\item{dt}{time interval (of inverse of sampling freq.)}
}
\description{
Response calculation of SDOF using Nigam & Jennings algorithm
}
