% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parzen.R
\name{parzen_window}
\alias{parzen_window}
\title{Parzen window smoothing of power spectral densities}
\usage{
parzen_window(freq, u)
}
\arguments{
\item{freq}{frequency}

\item{u, bandwidth}{param of window.
these two are related with each other as
'u <- 280/bandwidth/151'}

\item{power}{power spectral density (PSD)}

\item{dat1, dat2}{time-series data}

\item{dt}{time interval}

\item{smooth}{if FALSE, it returns raw periodgram}
}
\description{
Parzen window smoothing of power spectral densities
}
