\name{resp-package}
\alias{resp-package}
\alias{resp}
\docType{package}
\title{
\packageTitle{resp}
}
\description{
\packageDescription{resp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{resp}
\packageIndices{resp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{resp}

Maintainer: \packageMaintainer{resp}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
